pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "workspace"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "mypy>=1.17.1",
    "nbconvert>=7.16.6",
    "networkx>=3.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.7",
    "scikit-learn>=1.7.1",
    "sqlglot>=27.6.0",
]

[tool.ruff]
target-version = "py313"
line-length = 79

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    "UP",
    "PD",
    "NPY"
]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 7

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-config=.coveragerc --cov-report=html"
pythonpath = ["src"]
testpaths = ["tests"]
